New Update

+++ database/migrations/2024_01_30_022118_create_karyawans_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama', 100);
             $table->string('email' );
             $table->string('alamat');
-            $table->integer('no_telp');
+            $table->bigInteger('no_telp');
             $table->timestamps();
         });
     }
